+++ helper/parser/parser.c
@@ -116,7 +116,7 @@ void check_where(char* token, Query** query){
 }
 
 bool exceed_max_len(char* token, Query** query, int max_len, char* current_str){
-    if (strlen(token)>max_len){
+    if ((int)strlen(token)>max_len){
         (*query)->cmd_type = INVALID;
         fprintf(stderr, "Syntax error: %d characters maximum allowed for %s.\n\n", max_len-1, current_str);
         return true;
@@ -131,7 +131,7 @@ bool is_valid_identifier(char* token, Query** query){
         "FROM", "INTO", "WHERE", "JOIN", "ON", "INT", "STRING", "DOUBLE"};
     int i;
 
-    for(i=0; i<sizeof(banned_name_list) / sizeof(banned_name_list[0]); i++){
+    for(i=0; i<(int)(sizeof(banned_name_list) / sizeof(banned_name_list[0])); i++){
         if(strcasecmp(token, banned_name_list[i]) == 0){
             (*query)->cmd_type = INVALID;
             fprintf(stderr, "Syntax error: '%s' is a reserved keyword.\n\n", token);
@@ -139,7 +139,7 @@ bool is_valid_identifier(char* token, Query** query){
         }
     }
     // check no special character allowed
-    for(i=0; i<strlen(token); i++){
+    for(i=0; i<(int)strlen(token); i++){
         if(!isalnum(token[i]) && token[i] != '_'){
             (*query)->cmd_type = INVALID;
             fprintf(stderr, "Syntax error: special character '%c' is not allowed.\n\n", token[i]);
